+++ src/main/java/com/pbs/ams/web/model/AmsTradeOrder.java
@@ -9,7 +9,7 @@ public class AmsTradeOrder implements Serializable {
      *
      * @mbg.generated
      */
-    private Integer tradeOrderId;
+    private Long tradeOrderId;
 
     /**
      * 合约编号
@@ -86,15 +86,15 @@ public class AmsTradeOrder implements Serializable {
      *
      * @mbg.generated
      */
-    private Integer operatorId;
+    private Long operatorId;
 
     private static final long serialVersionUID = 1L;
 
-    public Integer getTradeOrderId() {
+    public Long getTradeOrderId() {
         return tradeOrderId;
     }
 
-    public void setTradeOrderId(Integer tradeOrderId) {
+    public void setTradeOrderId(Long tradeOrderId) {
         this.tradeOrderId = tradeOrderId;
     }
 
@@ -178,11 +178,11 @@ public class AmsTradeOrder implements Serializable {
         this.updateTime = updateTime;
     }
 
-    public Integer getOperatorId() {
+    public Long getOperatorId() {
         return operatorId;
     }
 
-    public void setOperatorId(Integer operatorId) {
+    public void setOperatorId(Long operatorId) {
         this.operatorId = operatorId;
     }
 }
