+++ src/main/java/com/pbs/ams/web/model/AmsTradeOrderFeedSnaps.java
@@ -16,7 +16,7 @@ public class AmsTradeOrderFeedSnaps implements Serializable {
      *
      * @mbg.generated
      */
-    private Integer tradeOrderFeedId;
+    private Long tradeOrderFeedId;
 
     /**
      * 合同编号
@@ -72,7 +72,7 @@ public class AmsTradeOrderFeedSnaps implements Serializable {
      *
      * @mbg.generated
      */
-    private Integer operatorId;
+    private Long operatorId;
 
     private static final long serialVersionUID = 1L;
 
@@ -84,11 +84,11 @@ public class AmsTradeOrderFeedSnaps implements Serializable {
         this.snapsTime = snapsTime;
     }
 
-    public Integer getTradeOrderFeedId() {
+    public Long getTradeOrderFeedId() {
         return tradeOrderFeedId;
     }
 
-    public void setTradeOrderFeedId(Integer tradeOrderFeedId) {
+    public void setTradeOrderFeedId(Long tradeOrderFeedId) {
         this.tradeOrderFeedId = tradeOrderFeedId;
     }
 
@@ -148,11 +148,11 @@ public class AmsTradeOrderFeedSnaps implements Serializable {
         this.fee = fee;
     }
 
-    public Integer getOperatorId() {
+    public Long getOperatorId() {
         return operatorId;
     }
 
-    public void setOperatorId(Integer operatorId) {
+    public void setOperatorId(Long operatorId) {
         this.operatorId = operatorId;
     }
 }
